# specs/001-app-shell-dashboard/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CleverDash API — 001 Ogrodje in dashboard
  version: 1.0.0
  description: |
    Pogodba funkcionalnosti 001. Vsaka zmožnost je najprej endpoint, šele nato zaslon
    (člen III ustave). Poti so angleške, kebab-case in verzionirane; domenska besedila
    (naslovi zavihkov, sporočila) so slovenska (člen X).

    Enotni izvor: API je na isti domeni kot aplikacija, pod `/api/v1`. CORS ni
    konfiguriran in ni potreben (člen II).

    Glava `Idempotency-Key` je sprejeta na vseh mutacijah razen `POST /auth/login` in
    `POST /auth/refresh` — utemeljitev je v plan.md, razdelek Complexity Tracking.
servers:
  - url: https://app.si/api/v1
    description: Produkcija
  - url: http://localhost:3000/api/v1
    description: Razvoj (dev-server proxy pošlje /api sem)
...
security:
  - bearerAuth: []
paths:
...
  /api-keys:
    get:
      tags: [api-keys]
      summary: Seznam ključev za avtomatizacijo
      description: Čistopis ključa se ne vrne nikoli; v seznamu je samo predpona.
      responses:
        '200':
          description: Ključi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
    post:
      tags: [api-keys]
      summary: Ustvari ključ z omejenim obsegom
      description: |
        Čistopis se pokaže **samo v tem odgovoru** in ni obnovljiv. Ključ brez obsegov ni
        dovoljen (člen III).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, scopes]
              properties:
                label:
                  type: string
                scopes:
                  type: array
                  minItems: 1
                  items:
                    type: string
                expiresAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Ustvarjen
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiKey'
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: Čistopis ključa, prikazan samo enkrat
        '400':
          $ref: '#/components/responses/BadRequest'
  /api-keys/{keyId}:
    delete:
      tags: [api-keys]
      summary: Prekliči ključ
      parameters:
        - name: keyId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Preklican
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Kratkoživ dostopni žeton (15 min) za web in Android
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Avtentikacija za avtomatizacijo (n8n). Ključ ima omejen obseg; veljaven ključ sam
        po sebi ne pomeni administratorskih pravic (FR-013, člen III).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Ponovljen klic z isto vrednostjo vrne prvotni rezultat. Ista vrednost z drugačnim
        telesom zahteve vrne `422`.
      schema:
        type: string
        maxLength: 200
  responses:
    BadRequest:
      description: Neveljavna zahteva
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Ni najdeno
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457. Sporočilo je v slovenščini in brez tehničnih podrobnosti (FR-026).
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        correlationId:
          type: string
          description: ID korelacije iz dnevnika, da je prijavljeno napako mogoče najti
...
    ApiKey:
      type: object
      required: [id, label, keyPrefix, scopes]
      properties:
        id:
          type: string
        label:
          type: string
        keyPrefix:
          type: string
          description: Prvih 8 znakov, da je ključ prepoznaven brez razkritja
        scopes:
          type: array
          items:
            type: string
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
...